+++ src/presentation/constants/navigation/authStackNavigation.tsx
@@ -1,0 +1,23 @@
+import React from 'react';
+import {createNativeStackNavigator} from '@react-navigation/native-stack';
+import LoginScreen from '@src/presentation/screens/auth/login';
+import RegisterScreen from '@src/presentation/screens/auth/register';
+
+const Stack = createNativeStackNavigator();
+
+const AuthStackNavigation = () => (
+  <Stack.Navigator>
+    <Stack.Screen
+      name="Login"
+      component={LoginScreen}
+      options={{headerShown: false}}
+    />
+    <Stack.Screen
+      name="Register"
+      component={RegisterScreen}
+      options={{headerShown: false}}
+    />
+  </Stack.Navigator>
+);
+
+export default AuthStackNavigation;
