+++ src/presentation/screens/goal/goalDetailsScreen.tsx
@@ -4,6 +4,7 @@ import {BaseLayout} from '../../constants/layouts';
 import {ActivityCard, GoalDetailCard} from '../../components/cards';
 import {P} from '../../constants/text';
 import {UpdateActionSheet} from '../actionSheets';
+import ScreenHeader from '@src/presentation/components/screenHeader';
 
 const GoalDetailsScreen = () => {
   const [open, setOpen] = useState(false);
@@ -12,28 +13,30 @@ const GoalDetailsScreen = () => {
     setOpen(!open);
   };
   return (
-    <BaseLayout>
-      <View>
+    <>
+      <BaseLayout>
         <View>
-          <GoalDetailCard />
-          <View className="mt-6 pt-3">
-            <P className="mb-3">Recent Updates</P>
-            <View className="h-[400px]">
-              <ScrollView>
-                {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((card, i) => (
-                  <ActivityCard onPress={() => null} key={i} />
-                ))}
-              </ScrollView>
+          <View>
+            <GoalDetailCard />
+            <View className="mt-6 pt-3">
+              <P className="mb-3">Recent Updates</P>
+              <View className="h-[400px]">
+                <ScrollView>
+                  {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((card, i) => (
+                    <ActivityCard onPress={() => null} key={i} />
+                  ))}
+                </ScrollView>
+              </View>
             </View>
           </View>
         </View>
-      </View>
-      <UpdateActionSheet
-        open={open}
-        toggleOverlay={handleToggleOverlay}
-        setOpen={setOpen}
-      />
-    </BaseLayout>
+        <UpdateActionSheet
+          open={open}
+          toggleOverlay={handleToggleOverlay}
+          setOpen={setOpen}
+        />
+      </BaseLayout>
+    </>
   );
 };
